requirements.txt
pandas
scikit-learn
matplotlib
Flask-SQLAlchemy
